+++ CHANGES.rst
@@ -23,8 +23,8 @@ TODOs
 * wltp_db: Improve test-metrics with group-by classes/phases.
 * wltp_db: Add diagrams auto-generated from test-metrics into generated site (Getting Involved section).
 * wltp_db: Speed-up tests by caching files to read and compare.
-* core: model: Improve model-preprocessing by interleaving "octapus" merging stacked-models
-  between 2 validation stages.
+* core: model: Enhance model-preprocessing by interleaving "octapus" merging stacked-models
+  between validation stages.
 * core: Move calculations as class-methods to provide for overriding certain parts of the algorithm.
 * core: Support to provide and override arbitrary model-data, and ask for arbitrary output-ones
   by topologically sorting the  graphs of the calculation-dependencies.
@@ -42,7 +42,7 @@ v0.0.9-beta, XX-Aug-2014
 * Possible to overide WLTP-Class, Target-V & Slope, Gears and other "results".
 * Include sample-vehicles along with the distribution.
 * build: Use TravisCI as integration server, Coveralls.io as test-coverage service-providers.
-* build: Not more possible to distribute it as .EXE; need proper python envronment.
+* build: Not possible anymore to distribute it as .EXE; need a proper python envronment.
 
 v0.0.8-alpha, 04-Aug-2014
 ---------------------------------------